+++ src/app/landing/components/HeroSection.tsx
